Move or rename D8-related items
